Update to v12.0
Major overhaul of system button logic. Instead of a series of menus, a single, larger menu pane was added.
Search field added which filters available system buttons based on comparison with what is in the search field.
When toggling the menu buttons, the map now is hidden to prevent several issues. Clicking outside the menu closes it and returns the map to visible.