System Map updates
Added scrollwheel zooming tied to the slider bar.
Changed zoom logic so it zooms towards the center instead of top center. (TODO: zoom towards center of viewport instead)
Fixed bug where if you click and drag over the location of a system, it would go to that system. Now that is disabled if the user was dragging the map.